test(ui): Move helper logic tests out of source files
