+++ ui/src/helpers/handle-multisort.ts
@@ -92,79 +92,3 @@ export const convertToBackendSorting = (
     .map((sort) => (sort.desc ? `-${sort.id}` : `${sort.id}`))
     .join(',');
 };
-
-//
-// Unit tests for multisorting
-//
-
-if (import.meta.vitest) {
-  const { it, expect } = import.meta.vitest;
-
-  // Test the updateColumnSorting() function
-
-  it('adds a new column to an empty sorting state', () => {
-    const columns = undefined;
-    const column = { id: 'column1', desc: false };
-
-    expect(updateColumnSorting(columns, column)).toEqual([
-      { id: 'column1', desc: false },
-    ]);
-  });
-
-  it('adds a new column to an existing sorting state', () => {
-    const columns = [{ id: 'column1', desc: false }];
-    const column = { id: 'column2', desc: false };
-
-    expect(updateColumnSorting(columns, column)).toEqual([
-      { id: 'column1', desc: false },
-      { id: 'column2', desc: false },
-    ]);
-  });
-
-  it('updates the sorting state of an existing column from desc -> asc', () => {
-    const columns = [{ id: 'column1', desc: false }];
-    const column = { id: 'column1', desc: false };
-
-    expect(updateColumnSorting(columns, column)).toEqual([
-      { id: 'column1', desc: true },
-    ]);
-  });
-
-  it('when an existing column is asc -> removes the column from the sorting state', () => {
-    const columns = [{ id: 'column1', desc: true }];
-    const column = { id: 'column1', desc: true };
-
-    expect(updateColumnSorting(columns, column)).toEqual(undefined);
-  });
-
-  it('more complicated sorting state update', () => {
-    const columns = [
-      { id: 'column1', desc: false },
-      { id: 'column2', desc: true },
-      { id: 'column3', desc: false },
-    ];
-    const column = { id: 'column2', desc: true };
-
-    expect(updateColumnSorting(columns, column)).toEqual([
-      { id: 'column1', desc: false },
-      { id: 'column3', desc: false },
-    ]);
-  });
-
-  // Test the convertToBackendSorting() function
-
-  it('converts sorting state to backend format', () => {
-    const sorting: SortingState = [
-      { id: 'column1', desc: false },
-      { id: 'column2', desc: true },
-    ];
-
-    expect(convertToBackendSorting(sorting)).toEqual('column1,-column2');
-  });
-
-  it('returns undefined for empty sorting state', () => {
-    const sorting: SortingState = [];
-
-    expect(convertToBackendSorting(sorting)).toEqual(undefined);
-  });
-}
