+++ ui/src/helpers/licenses/spdx-expression.ts
@@ -161,199 +161,3 @@ export function parseLicenseExpression(
 export function isAtomicLicenseExpression(expression: string): boolean {
   return parseLicenseExpression(expression).kind === 'atomic';
 }
-
-//
-// Unit tests for SPDX expression parsing helpers.
-//
-
-if (import.meta.vitest) {
-  const { describe, it, expect } = import.meta.vitest;
-
-  describe('normalizeLicenseExpression', () => {
-    it('normalizes valid SPDX expressions', () => {
-      expect(normalizeLicenseExpression('  mit or apache-2.0  ')).toBe(
-        'Apache-2.0 OR MIT'
-      );
-    });
-
-    it('normalizes deprecated GPL plus syntax', () => {
-      expect(normalizeLicenseExpression('GPL-2.0+')).toBe('GPL-2.0-or-later');
-    });
-
-    it('returns a trimmed original string for invalid expressions', () => {
-      expect(normalizeLicenseExpression(' MIT OR ( ')).toBe('MIT OR (');
-    });
-  });
-
-  describe('parseLicenseExpression', () => {
-    it('parses a single SPDX license identifier as an atomic expression', () => {
-      expect(parseLicenseExpression('MIT')).toEqual({
-        kind: 'atomic',
-        rawExpression: 'MIT',
-        normalizedExpression: 'MIT',
-        node: {
-          kind: 'license',
-          license: 'MIT',
-          exception: undefined,
-        },
-      });
-    });
-
-    it('parses SPDX expressions with AND / OR into conjunction nodes', () => {
-      expect(parseLicenseExpression('MIT AND Apache-2.0')).toEqual({
-        kind: 'compound',
-        rawExpression: 'MIT AND Apache-2.0',
-        normalizedExpression: 'Apache-2.0 AND MIT',
-        node: {
-          kind: 'conjunction',
-          conjunction: 'and',
-          left: {
-            kind: 'license',
-            license: 'MIT',
-            exception: undefined,
-          },
-          right: {
-            kind: 'license',
-            license: 'Apache-2.0',
-            exception: undefined,
-          },
-        },
-      });
-    });
-
-    it('parses nested SPDX expressions recursively', () => {
-      expect(
-        parseLicenseExpression('MIT AND (Apache-2.0 OR BSD-3-Clause)')
-      ).toEqual({
-        kind: 'compound',
-        rawExpression: 'MIT AND (Apache-2.0 OR BSD-3-Clause)',
-        normalizedExpression: 'MIT AND (Apache-2.0 OR BSD-3-Clause)',
-        node: {
-          kind: 'conjunction',
-          conjunction: 'and',
-          left: {
-            kind: 'license',
-            license: 'MIT',
-            exception: undefined,
-          },
-          right: {
-            kind: 'conjunction',
-            conjunction: 'or',
-            left: {
-              kind: 'license',
-              license: 'Apache-2.0',
-              exception: undefined,
-            },
-            right: {
-              kind: 'license',
-              license: 'BSD-3-Clause',
-              exception: undefined,
-            },
-          },
-        },
-      });
-    });
-
-    it('parses WITH exceptions as atomic license nodes', () => {
-      expect(
-        parseLicenseExpression('GPL-2.0-only WITH Classpath-exception-2.0')
-      ).toEqual({
-        kind: 'atomic',
-        rawExpression: 'GPL-2.0-only WITH Classpath-exception-2.0',
-        normalizedExpression: 'GPL-2.0-only WITH Classpath-exception-2.0',
-        node: {
-          kind: 'license',
-          license: 'GPL-2.0-only',
-          exception: 'Classpath-exception-2.0',
-        },
-      });
-    });
-
-    it('upgrades deprecated GPL plus syntax during parsing', () => {
-      expect(parseLicenseExpression('GPL-2.0+')).toEqual({
-        kind: 'atomic',
-        rawExpression: 'GPL-2.0+',
-        normalizedExpression: 'GPL-2.0-or-later',
-        node: {
-          kind: 'license',
-          license: 'GPL-2.0-or-later',
-          exception: undefined,
-        },
-      });
-    });
-
-    it('accepts lower-case operators and normalizes them', () => {
-      expect(parseLicenseExpression('mit or apache-2.0')).toEqual({
-        kind: 'compound',
-        rawExpression: 'mit or apache-2.0',
-        normalizedExpression: 'Apache-2.0 OR MIT',
-        node: {
-          kind: 'conjunction',
-          conjunction: 'or',
-          left: {
-            kind: 'license',
-            license: 'MIT',
-            exception: undefined,
-          },
-          right: {
-            kind: 'license',
-            license: 'Apache-2.0',
-            exception: undefined,
-          },
-        },
-      });
-    });
-
-    it('supports custom LicenseRef identifiers', () => {
-      expect(parseLicenseExpression('DocumentRef-foo:LicenseRef-bar')).toEqual({
-        kind: 'atomic',
-        rawExpression: 'DocumentRef-foo:LicenseRef-bar',
-        normalizedExpression: 'DocumentRef-foo:LicenseRef-bar',
-        node: {
-          kind: 'license-ref',
-          documentRef: 'DocumentRef-foo',
-          licenseRef: 'LicenseRef-bar',
-        },
-      });
-    });
-
-    it('returns an invalid result instead of throwing for bad input', () => {
-      const result = parseLicenseExpression('MIT OR (');
-
-      expect(result.kind).toBe('invalid');
-      expect(result.rawExpression).toBe('MIT OR (');
-      expect(result.normalizedExpression).toBe('MIT OR (');
-
-      if (result.kind === 'invalid') {
-        expect(result.error.length).toBeGreaterThan(0);
-      }
-    });
-
-    it('returns an invalid result for an empty expression', () => {
-      expect(parseLicenseExpression('   ')).toEqual({
-        kind: 'invalid',
-        rawExpression: '   ',
-        normalizedExpression: '',
-        error: 'License expression is empty.',
-      });
-    });
-  });
-
-  describe('isAtomicLicenseExpression', () => {
-    it('returns true for a single SPDX license', () => {
-      expect(isAtomicLicenseExpression('Apache-2.0')).toBe(true);
-    });
-
-    it('returns false for compound expressions', () => {
-      expect(isAtomicLicenseExpression('Apache-2.0 OR MIT')).toBe(false);
-    });
-
-    it('returns false for invalid expressions', () => {
-      expect(isAtomicLicenseExpression('MIT OR (')).toBe(false);
-    });
-
-    it('returns true for normalized GPL plus syntax because it remains atomic', () => {
-      expect(isAtomicLicenseExpression('GPL-2.0+')).toBe(true);
-    });
-  });
-}
